soft_repo.gl: Rename the release/* branches to distro/*.
The rationale here is to make it more apparent to maintainers that the
branches are not per-local-release version, but for major distro
releases should any updates need to be pushed to stable releases.

Due to the history, we will import software with previous Mandriva and
MES distro branches, which the import script helpfully renames for us to
distro/mdv* and distro/mes* appropriately. For our own distro branches
in the future, we should name them distro/mga*